Control.Monad.Random.Lazy

Description

Random monads that are lazy in the generator state. For a strict version, see Control.Monad.Random.Strict, which has the same interface.

The Rand monad

type Rand g = RandT g Identity

A random monad parameterized by the type g of the generator to carry.

The return function leaves the generator unchanged, while >>= uses the final generator of the first computation as the initial generator of the second.

liftRand

Arguments

:: (g -> (a, g))
pure random transformer

-> Rand g a
equivalent generator-passing computation

Construct a random monad computation from a function. (The inverse of runRand.)

runRand

Arguments

:: Rand g a
generator-passing computation to execute

-> g
initial generator

-> (a, g)
return value and final generator

Unwrap a random monad computation as a function. (The inverse of liftRand.)

evalRand

Arguments

:: Rand g a
generator-passing computation to execute

-> g
initial generator

-> a
return value of the random computation

Evaluate a random computation with the given initial generator and return the final value, discarding the final generator.

evalRand m s = fst (runRand m s)

execRand

Arguments

:: Rand g a
generator-passing computation to execute

-> g
initial generator

-> g
final generator

Evaluate a random computation with the given initial generator and return the final generator, discarding the final value.

execRand m s = snd (runRand m s)

mapRand :: ((a, g) -> (b, g)) -> Rand g a -> Rand g b

Map both the return value and final generator of a computation using the given function.

runRand (mapRand f m) = f . runRand m

withRand :: (g -> g) -> Rand g a -> Rand g a

withRand f m executes action m on a generator modified by applying f.

withRand f m = modify f >> m

evalRandIO :: Rand StdGen a -> IO a

Evaluate a random computation in the IO monad, splitting the global standard generator to get a new one for the computation.

The RandT monad transformer

data RandT g (m :: Type -> Type) a

A random transformer monad parameterized by:

g - The generator.

m - The inner monad.

The return function leaves the generator unchanged, while >>= uses the final generator of the first computation as the initial generator of the second.

liftRandT

Arguments

:: (g -> m (a, g))
impure random transformer

-> RandT g m a
equivalent generator-passing computation

Construct a random monad computation from an impure function. (The inverse of runRandT.)

runRandT

Arguments

:: RandT g m a
generator-passing computation to execute

-> g
initial generator

-> m (a, g)
return value and final generator

Unwrap a random monad computation as an impure function. (The inverse of liftRandT.)

evalRandT :: Monad m => RandT g m a -> g -> m a

Evaluate a random computation with the given initial generator and return the final value, discarding the final generator.

evalRandT m g = liftM fst (runRandT m g)

execRandT :: Monad m => RandT g m a -> g -> m g

Evaluate a random computation with the given initial generator and return the final generator, discarding the final value.

execRandT m g = liftM snd (runRandT m g)

mapRandT :: (m (a, g) -> n (b, g)) -> RandT g m a -> RandT g n b

Map both the return value and final generator of a computation using the given function.

runRandT (mapRandT f m) = f . runRandT m

withRandT :: forall g (m :: Type -> Type) a. (g -> g) -> RandT g m a -> RandT g m a

withRandT f m executes action m on a generator modified by applying f.

withRandT f m = modify f >> m

evalRandTIO :: MonadIO m => RandT StdGen m a -> m a

Evaluate a random computation that is embedded in the IO monad, splitting the global standard generator to get a new one for the computation.

Some convenience re-exports

class Random a where

The class of types for which random values can be generated. Most instances of Random will produce values that are uniformly distributed on the full range, but for those types without a well-defined "full range" some sensible default subrange will be selected.

Random exists primarily for backwards compatibility with version 1.1 of this library. In new code, use the better specified Uniform and UniformRange instead.

Since: random-1.0.0

Minimal complete definition

Nothing

Methods

randomR :: RandomGen g => (a, a) -> g -> (a, g)

Takes a range (lo,hi) and a pseudo-random number generator g, and returns a pseudo-random value uniformly distributed over the closed interval [lo,hi], together with a new generator. It is unspecified what happens if lo>hi, but usually the values will simply get swapped.

>>> let gen = mkStdGen 2021
>>> fst $ randomR ('a', 'z') gen
't'
>>> fst $ randomR ('z', 'a') gen
't'

For continuous types there is no requirement that the values lo and hi are ever produced, but they may be, depending on the implementation and the interval.

There is no requirement to follow the Ord instance and the concept of range can be defined on per type basis. For example product types will treat their values independently:

>>> fst $ randomR (('a', 5.0), ('z', 10.0)) $ mkStdGen 2021
('t',6.240232662366563)

In case when a lawful range is desired uniformR should be used instead.

Since: random-1.0.0

default randomR :: (RandomGen g, UniformRange a) => (a, a) -> g -> (a, g)

random :: RandomGen g => g -> (a, g)

The same as randomR, but using a default range determined by the type:

For bounded types (instances of Bounded, such as Char), the range is normally the whole type.

For floating point types, the range is normally the closed interval [0,1].

For Integer, the range is (arbitrarily) the range of Int.

Since: random-1.0.0

default random :: (RandomGen g, Uniform a) => g -> (a, g)

randomRs :: RandomGen g => (a, a) -> g -> [a]

Plural variant of randomR, producing an infinite list of pseudo-random values instead of returning a new generator.

Since: random-1.0.0

randoms :: RandomGen g => g -> [a]

Plural variant of random, producing an infinite list of pseudo-random values instead of returning a new generator.

Since: random-1.0.0

class Uniform a

The class of types for which a uniformly distributed value can be drawn from all possible values of the type.

Since: random-1.2.0

class Finite a

A type class for data with a finite number of inhabitants. This type class is used in default implementations of Uniform.

Users are not supposed to write instances of Finite manually. There is a default implementation in terms of Generic instead.

>>> :set -XDeriveGeneric -XDeriveAnyClass
>>> import GHC.Generics (Generic)
>>> data MyBool = MyTrue | MyFalse deriving (Generic, Finite)
>>> data Action = Code MyBool | Eat (Maybe Bool) | Sleep deriving (Generic, Finite)

class UniformRange a

The class of types for which a uniformly distributed value can be drawn from a range.

Since: random-1.2.0

Minimal complete definition

uniformRM

data StdGen

The standard pseudo-random number generator.

class RandomGen g where

RandomGen is an interface to pure pseudo-random number generators.

StdGen is the standard RandomGen instance provided by this library.

Since: random-1.0.0

Minimal complete definition

split, (genWord32 | genWord64 | next, genRange)

Methods

next :: g -> (Int, g)

Returns an Int that is uniformly distributed over the range returned by genRange (including both end points), and a new generator. Using next is inefficient as all operations go via Integer. See here for more details. It is thus deprecated.

Since: random-1.0.0

genWord8 :: g -> (Word8, g)

Returns a Word8 that is uniformly distributed over the entire Word8 range.

Since: random-1.2.0

genWord16 :: g -> (Word16, g)

Returns a Word16 that is uniformly distributed over the entire Word16 range.

Since: random-1.2.0

genWord32 :: g -> (Word32, g)

Returns a Word32 that is uniformly distributed over the entire Word32 range.

Since: random-1.2.0

genWord64 :: g -> (Word64, g)

Returns a Word64 that is uniformly distributed over the entire Word64 range.

Since: random-1.2.0

genWord32R :: Word32 -> g -> (Word32, g)

genWord32R upperBound g returns a Word32 that is uniformly distributed over the range [0, upperBound].

Since: random-1.2.0

genWord64R :: Word64 -> g -> (Word64, g)

genWord64R upperBound g returns a Word64 that is uniformly distributed over the range [0, upperBound].

Since: random-1.2.0

genShortByteString :: Int -> g -> (ShortByteString, g)

genShortByteString n g returns a ShortByteString of length n filled with pseudo-random bytes.

Since: random-1.2.0

genRange :: g -> (Int, Int)

Yields the range of values returned by next.

It is required that:

If (a, b) = genRange g, then a < b.

genRange must not examine its argument so the value it returns is determined only by the instance of RandomGen.

The default definition spans the full range of Int.

Since: random-1.0.0

split :: g -> (g, g)

Returns two distinct pseudo-random number generators.

Implementations should take care to ensure that the resulting generators are not correlated. Some pseudo-random number generators are not splittable. In that case, the split implementation should fail with a descriptive error message.

Since: random-1.0.0

mkStdGen :: Int -> StdGen

Constructs a StdGen deterministically.

genByteString :: RandomGen g => Int -> g -> (ByteString, g)

Generates a ByteString of the specified size using a pure pseudo-random number generator. See uniformByteStringM for the monadic version.

Examples

Expand

>>> import System.Random
>>> import Data.ByteString
>>> let pureGen = mkStdGen 137
>>> unpack . fst . genByteString 10 $ pureGen
[51,123,251,37,49,167,90,109,1,4]

Since: random-1.2.0

initStdGen :: MonadIO m => m StdGen

Initialize StdGen using system entropy (i.e. /dev/urandom) when it is available, while falling back on using system time as the seed.

Since: random-1.2.1

setStdGen :: MonadIO m => StdGen -> m ()

Sets the global pseudo-random number generator. Overwrites the contents of globalStdGen

Since: random-1.0.0

getStdGen :: MonadIO m => m StdGen

Gets the global pseudo-random number generator. Extracts the contents of globalStdGen

Since: random-1.0.0

newStdGen :: MonadIO m => m StdGen

Applies split to the current global pseudo-random generator globalStdGen, updates it with one of the results, and returns the other.

Since: random-1.0.0

getStdRandom :: MonadIO m => (StdGen -> (a, StdGen)) -> m a

Uses the supplied function to get a value from the current global random generator, and updates the global generator with the new generator returned by the function. For example, rollDice produces a pseudo-random integer between 1 and 6:

>>> rollDice = getStdRandom (randomR (1, 6))
>>> replicateM 10 (rollDice :: IO Int)
[5,6,6,1,1,6,4,2,4,1]

This is an outdated function and it is recommended to switch to its equivalent applyAtomicGen instead, possibly with the globalStdGen if relying on the global state is acceptable.

>>> import System.Random.Stateful
>>> rollDice = applyAtomicGen (uniformR (1, 6)) globalStdGen
>>> replicateM 10 (rollDice :: IO Int)
[4,6,1,1,4,4,3,2,1,2]

Since: random-1.0.0

randomRIO :: (Random a, MonadIO m) => (a, a) -> m a

A variant of randomRM that uses the global pseudo-random number generator globalStdGen

>>> randomRIO (2020, 2100) :: IO Int
2040

Similar to randomIO, this function is equivalent to getStdRandom randomR and is included in this interface for historical reasons and backwards compatibility. It is recommended to use uniformRM instead, possibly with the globalStdGen if relying on the global state is acceptable.

>>> import System.Random.Stateful
>>> uniformRM (2020, 2100) globalStdGen :: IO Int
2079

Since: random-1.0.0

randomIO :: (Random a, MonadIO m) => m a

A variant of randomM that uses the global pseudo-random number generator globalStdGen.

>>> import Data.Int
>>> randomIO :: IO Int32
-1580093805

This function is equivalent to getStdRandom random and is included in this interface for historical reasons and backwards compatibility. It is recommended to use uniformM instead, possibly with the globalStdGen if relying on the global state is acceptable.

>>> import System.Random.Stateful
>>> uniformM globalStdGen :: IO Int32
-1649127057

Since: random-1.0.0

module Control.Monad.Random.Class

class Functor (f :: Type -> Type) where

Minimal complete definition

fmap

Methods

fmap :: (a -> b) -> f a -> f b

(<$) :: a -> f b -> f a

class Applicative m => Monad (m :: Type -> Type) where

Minimal complete definition

(>>=)

Methods

(>>=) :: m a -> (a -> m b) -> m b

(>>) :: m a -> m b -> m b

return :: a -> m a

join :: Monad m => m (m a) -> m a

liftM :: Monad m => (a1 -> r) -> m a1 -> m r

class (Alternative m, Monad m) => MonadPlus (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

mzero :: m a

mplus :: m a -> m a -> m a

class Monad m => MonadFail (m :: Type -> Type) where

Methods

fail :: String -> m a

mapM_ :: (Foldable t, Monad m) => (a -> m b) -> t a -> m ()

sequence_ :: (Foldable t, Monad m) => t (m a) -> m ()

mapM :: (Traversable t, Monad m) => (a -> m b) -> t a -> m (t b)

sequence :: (Traversable t, Monad m) => t (m a) -> m (t a)

(=<<) :: Monad m => (a -> m b) -> m a -> m b

forM_ :: (Foldable t, Monad m) => t a -> (a -> m b) -> m ()

msum :: (Foldable t, MonadPlus m) => t (m a) -> m a

void :: Functor f => f a -> f ()

forM :: (Traversable t, Monad m) => t a -> (a -> m b) -> m (t b)

ap :: Monad m => m (a -> b) -> m a -> m b

liftM2 :: Monad m => (a1 -> a2 -> r) -> m a1 -> m a2 -> m r

liftM3 :: Monad m => (a1 -> a2 -> a3 -> r) -> m a1 -> m a2 -> m a3 -> m r

liftM4 :: Monad m => (a1 -> a2 -> a3 -> a4 -> r) -> m a1 -> m a2 -> m a3 -> m a4 -> m r

liftM5 :: Monad m => (a1 -> a2 -> a3 -> a4 -> a5 -> r) -> m a1 -> m a2 -> m a3 -> m a4 -> m a5 -> m r

when :: Applicative f => Bool -> f () -> f ()

(<$!>) :: Monad m => (a -> b) -> m a -> m b

(<=<) :: Monad m => (b -> m c) -> (a -> m b) -> a -> m c

(>=>) :: Monad m => (a -> m b) -> (b -> m c) -> a -> m c

foldM :: (Foldable t, Monad m) => (b -> a -> m b) -> b -> t a -> m b

foldM_ :: (Foldable t, Monad m) => (b -> a -> m b) -> b -> t a -> m ()

replicateM :: Applicative m => Int -> m a -> m [a]

replicateM_ :: Applicative m => Int -> m a -> m ()

unless :: Applicative f => Bool -> f () -> f ()

filterM :: Applicative m => (a -> m Bool) -> [a] -> m [a]

forever :: Applicative f => f a -> f b

mapAndUnzipM :: Applicative m => (a -> m (b, c)) -> [a] -> m ([b], [c])

mfilter :: MonadPlus m => (a -> Bool) -> m a -> m a

zipWithM :: Applicative m => (a -> b -> m c) -> [a] -> [b] -> m [c]

zipWithM_ :: Applicative m => (a -> b -> m c) -> [a] -> [b] -> m ()

guard :: Alternative f => Bool -> f ()

class Monad m => MonadFix (m :: Type -> Type) where

Methods

mfix :: (a -> m a) -> m a

fix :: (a -> a) -> a

class (forall (m :: Type -> Type). Monad m => Monad (t m)) => MonadTrans (t :: (Type -> Type) -> Type -> Type) where

Methods

lift :: Monad m => m a -> t m a

class Monad m => MonadIO (m :: Type -> Type) where

Methods

liftIO :: IO a -> m a