Control.Monad.Random.Class

Description

A type class for random number generation monads. See http://www.haskell.org/haskellwiki/NewMonads/MonadRandom for the original version of this code.

Instances of this type class include Rand and monads created using RandT.

Documentation

class Monad m => MonadRandom m where

An interface to random number generation monads.

Methods

getRandom :: Random a => m a

Return a randomly-selected value of type a. See random for details.

getRandoms :: Random a => m [a]

Return an infinite stream of random values of type a. See randoms for details.

getRandomR :: Random a => (a, a) -> m a

Return a randomly-selected value of type a in the range (lo,hi). See randomR for details.

getRandomRs :: Random a => (a, a) -> m [a]

Return an infinite stream of randomly-selected value of type a in the range (lo,hi). See randomRs for details.

class Monad m => MonadSplit s m | m -> s where

An interface to monads with splittable state (as most random number generation monads will have). The intention is that the getSplit action splits the state, returning one half of the result, and setting the new state to the other.

Methods

getSplit :: m s